FIX: Handle Redis failure during jobs (#19)
MiniScheduler was previously able to detect 'orphaned' jobs when the entire `MiniScheduler::Manager` failed (e.g. a process restart). However, if something went wrong with the Redis connection during a job, the worker thread would be unable to remove the "RUNNING" status, and the job would be stuck in that state until its next scheduled run.

This commit makes a few changes to fix this:
- give each thread its own id, and use an expiring redis key to track its lifetime
- when changing a job to the 'RUNNING' state, update its `current_owner` from the manager's id to the worker thread's id. This ensures that the existing orphan-detection logic will pick up the job if the thread dies.
- ensure worker threads die when there is an internal mini_scheduler exception which could cause inconsistent job state
- automatically start new worker threads when existing ones die